fix: roving
